+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Braintree Android SDK Release Notes
+  
+## unreleased
+
+* PayPal
+  * Add `recurringBillingDetails`, `recurringBillingPlanType`, and `amountBreakdown` properties to `PayPalCheckoutRequest`. Enables RBA metadata to be passed for the PayPal Checkout Vault with Purchase flow
 
 ## 5.17.0 (2025-09-09)